Fix low quality audio streaming on Bluetooth devices
iOS permits two Bluetooth streaming profiles: HFP and A2DP. HFP is
intended for use with hands-free devices, typically when audio input
from the user via a microphone is required, e.g. when speaking during
a phone call. The A2DP profile is used for streaming high-quality
audio to devices that support this profile, and is typically what is
required by this plugin.

If both profiles are enabled in an audio session and available on the
target Bluetooth device, the LOWER quality HFP profile is given priority
- see the note in A2DP link below. Since there is no requirement for
audio input when playing a playlist, the A2DP profile is used whenever
possible; the HFP profile is only enabled for iOS versions older than
10.0, for which the A2DP profile is unavailable.

Links:
https://developer.apple.com/documentation/avfoundation/avaudiosessioncategoryoptions/avaudiosessioncategoryoptionallowbluetootha2dp
https://developer.apple.com/documentation/avfoundation/avaudiosessioncategoryoptions?language=objc